requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==45.0.6
defusedxml==0.7.1
Deprecated==1.2.18
Django==5.1.3
django-allauth==65.10.0
django-celery-beat==2.7.0
django-timezone-field==7.1
dotenv==0.9.9
idna==3.10
jaconv==0.4.0
kombu==5.4.2
oauthlib==3.3.1
pillow==11.3.0
prompt_toolkit==3.0.50
pycparser==2.22
PyJWT==2.10.1
pykakasi==2.3.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python3-openid==3.2.0
redis==5.2.1
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
wrapt==1.17.2
